requirements.txt
# Extracción
selenium==4.20.0
webdriver-manager==4.0.1
requests==2.31.0

# Base de datos
mysql-connector-python==8.4.0

# Utilidades
python-dotenv==1.0.1
